feat: Compile GResource and update icons
This commit introduces the compilation of GResource files, allowing for
the embedding of application resources.

It also includes the addition of new alert and check icons and updates
the application to use these icons for verification status. The
application ID has been standardized for better consistency.
Additionally, the runtime version for Flatpak has been updated to '49'.